+++ zarrs_storage/src/storage_sync.rs
@@ -1,14 +1,11 @@
 use std::sync::Arc;
 
 use auto_impl::auto_impl;
-use itertools::Itertools;
-
-use crate::MaybeBytesIterator;
 
 use super::{
     byte_range::{ByteRange, ByteRangeIterator},
-    Bytes, MaybeBytes, MaybeSend, MaybeSync, StorageError, StoreKey, StoreKeyOffsetValue,
-    StoreKeys, StoreKeysPrefixes, StorePrefix, StorePrefixes,
+    Bytes, MaybeBytes, MaybeBytesIterator, MaybeSend, MaybeSync, OffsetBytesIterator, StorageError,
+    StoreKey, StoreKeys, StoreKeysPrefixes, StorePrefix, StorePrefixes,
 };
 
 /// Readable storage traits.
@@ -113,55 +110,26 @@ pub trait ListableStorageTraits: MaybeSend + MaybeSync {
 ///
 /// # Panics
 /// Panics if a key ends beyond `usize::MAX`.
-pub fn store_set_partial_values<T: ReadableWritableStorageTraits>(
+pub fn store_set_partial_many<T: ReadableWritableStorageTraits>(
     store: &T,
-    key_offset_values: &[StoreKeyOffsetValue],
-    // truncate: bool,
+    key: &StoreKey,
+    offset_values: OffsetBytesIterator,
 ) -> Result<(), StorageError> {
-    // Group by key
-    key_offset_values
-        .iter()
-        .chunk_by(|key_offset_value| key_offset_value.key())
-        .into_iter()
-        .map(|(key, group)| (key.clone(), group.into_iter().cloned().collect::<Vec<_>>()))
-        .try_for_each(|(key, group)| {
-            // Lock the store key
-            // let mutex = store.mutex(&key)?;
-            // let _lock = mutex.lock();
+    // Read the store key
+    let bytes_out = store.get(key)?.unwrap_or_default();
+    let mut bytes_out: bytes::BytesMut = bytes_out.into();
 
-            // Read the store key
-            let bytes = store.get(&key)?.unwrap_or_default();
-            let mut bytes = Vec::<u8>::from(bytes);
+    // Update the store key
+    for (offset, value) in offset_values {
+        let offset = usize::try_from(offset).unwrap();
+        if bytes_out.len() < offset + value.len() {
+            bytes_out.resize(offset + value.len(), 0);
+        }
+        bytes_out[offset..offset + value.len()].copy_from_slice(&value);
+    }
 
-            // Convert to a mutable vector of the required length
-            let end_max = group
-                .iter()
-                .map(|key_offset_value| {
-                    usize::try_from(
-                        key_offset_value.offset() + key_offset_value.value().len() as u64,
-                    )
-                    .unwrap()
-                })
-                .max()
-                .unwrap();
-            if bytes.len() < end_max {
-                bytes.resize_with(end_max, Default::default);
-            }
-            // else if truncate {
-            //     bytes.truncate(end_max);
-            // };
-
-            // Update the store key
-            for key_offset_value in group {
-                let start = usize::try_from(key_offset_value.offset()).unwrap();
-                bytes[start..start + key_offset_value.value().len()]
-                    .copy_from_slice(key_offset_value.value());
-            }
-
-            // Write the store key
-            store.set(&key, Bytes::from(bytes))
-        })?;
-    Ok(())
+    // Write the store key
+    store.set(key, bytes_out.freeze())
 }
 
 /// Writable storage traits.
@@ -173,13 +141,22 @@ pub trait WritableStorageTraits: MaybeSend + MaybeSync {
     /// Returns a [`StorageError`] on failure to store.
     fn set(&self, key: &StoreKey, value: Bytes) -> Result<(), StorageError>;
 
-    /// Store bytes according to a list of [`StoreKeyOffsetValue`].
+    /// Store bytes from an offset and value.
     ///
     /// # Errors
     /// Returns a [`StorageError`] on failure to store.
-    fn set_partial_values(
+    fn set_partial(&self, key: &StoreKey, offset: u64, value: Bytes) -> Result<(), StorageError> {
+        self.set_partial_many(key, Box::new([(offset, value)].into_iter()))
+    }
+
+    /// Store bytes from a [`OffsetBytesIterator`].
+    ///
+    /// # Errors
+    /// Returns a [`StorageError`] on failure to store.
+    fn set_partial_many(
         &self,
-        key_offset_values: &[StoreKeyOffsetValue],
+        key: &StoreKey,
+        offset_values: OffsetBytesIterator,
     ) -> Result<(), StorageError>;
 
     /// Erase a [`StoreKey`].